Move images to a CardGame component
